check for engines failing to boot

+++ launch.sh
@@ -1,4 +1,6 @@
 #!/bin/bash -e
 docker build -t battlechess -f Dockerfile .   # 1>&2
-docker run -i --rm battlechess bash -c "cd /tmp/ && git clone -q https://github.com/$1/battlechess.git && python3 /tmp/battlechess/engine.py"
+if [ ! -z $1 ]; then
+  docker run -i --rm battlechess bash -c "cd /tmp/ && git clone -q https://github.com/$1/battlechess.git && python3 /tmp/battlechess/engine.py"
+fi
